ARM: mm: avoid flushing exec anonymous pages with non-aliasing caches
Given that userspace can't rely on the kernel to flush pages that are
mapped as executable after having been written to by the user, we can
avoid flushing anonymous executable pages when we don't have an aliasing
data cache.

This patch skips the cache maintenance in __sync_icache_dcache for this
scenario, bringing us in-line with the arm64 implementation.

Cc: Albin Tonnerre <albin.tonnerre@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>